Self Image Magnification
This was designed to magnify images that showed in an existing web
platform.

Yes, when an image is magnified, it shows pixelation. But this is what
was requested. And yes, there are plenty of jquery/javascript
magnifiers out there, but the ones I found involved small images and
high res variations. But again, that was not requested.

It has the ability to set magnification amount, and size of view able
window.

Note that this is edge to edge for the image and the magnified portion.